feat: alterar tipo de user_id e viewed_moment_id para BIGINT na migração de views

+++ src/database/migrations/20240208020816-create-views.js
@@ -11,14 +11,14 @@ module.exports = {
         allowNull: false
       },
       user_id: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.BIGINT(),
         allowNull: false,
         references: {model: 'users', key: 'id'},
         onUpdate: 'CASCADE',
         onDelete: 'CASCADE'  
       },
       viewed_moment_id: {
-        type: Sequelize.INTEGER(),
+        type: Sequelize.BIGINT(),
         allowNull: false,
         references: {
           model: 'moments',
